Refactor: [M3-6522-selection-card] Initial commit: exports
