+++ packages/manager/src/components/SelectionCard/CardBase.tsx
@@ -1,91 +1,14 @@
 import * as React from 'react';
-import Grid from '@mui/material/Unstable_Grid2';
-import { styled } from '@mui/material/styles';
-import { SxProps } from '@mui/system';
+import {
+  CardBaseGrid,
+  CardBaseHeadings,
+  CardBaseHeading,
+  CardBaseIcon,
+  CardBaseSubheading,
+} from './CardBase.styles';
+import type { SxProps } from '@mui/system';
 
-const CardBaseGrid = styled(Grid, {
-  label: 'CardBaseGrid',
-})<Partial<Props>>(({ theme, ...props }) => ({
-  alignItems: 'center',
-  backgroundColor: props.checked ? theme.bg.lightBlue2 : theme.bg.offWhite,
-  border: `1px solid ${theme.bg.main}`,
-  borderColor: props.checked ? theme.palette.primary.main : undefined,
-  height: '100%',
-  margin: 0,
-  minHeight: 60,
-  padding: `0 ${theme.spacing(1)} !important`,
-  position: 'relative',
-  transition:
-    'background-color 225ms ease-in-out, border-color 225ms ease-in-out',
-  width: '100%',
-
-  '&:hover': {
-    backgroundColor: props.checked ? theme.bg.lightBlue2 : theme.bg.main,
-    borderColor: props.checked
-      ? theme.palette.primary.main
-      : theme.color.border2,
-  },
-
-  '&:before': {
-    content: '""',
-    display: 'block',
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    width: 5,
-    height: '100%',
-    backgroundColor: 'transparent',
-    transition: theme.transitions.create('backgroundColor'),
-  },
-}));
-
-const CardBaseIcon = styled(Grid, {
-  label: 'CardBaseIcon',
-})(() => ({
-  display: 'flex',
-  alignItems: 'flex-end',
-  justifyContent: 'flex-end',
-  '& svg, & span': {
-    fontSize: 32,
-    color: '#939598',
-  },
-  '& img': {
-    maxHeight: 32,
-    maxWidth: 32,
-  },
-}));
-
-const CardBaseHeadings = styled(Grid, {
-  label: 'CardBaseHeadings',
-})(() => ({
-  flex: 1,
-  flexDirection: 'column',
-  justifyContent: 'space-around',
-  '& > div': {
-    lineHeight: 1.3,
-  },
-}));
-
-const CardBaseHeading = styled('div', {
-  label: 'CardBaseHeading',
-})(({ theme }) => ({
-  fontFamily: theme.font.bold,
-  fontSize: '1rem',
-  color: theme.color.headline,
-  wordBreak: 'break-word',
-  display: 'flex',
-  alignItems: 'center',
-  columnGap: theme.spacing(2),
-}));
-
-const CardBaseSubheading = styled('div', {
-  label: 'CardBaseSubheading',
-})(({ theme }) => ({
-  color: theme.palette.text.primary,
-  fontSize: '0.875rem',
-}));
-
-interface Props {
+export interface CardBaseProps {
   checked?: boolean;
   heading: string | JSX.Element;
   headingDecoration?: JSX.Element;
@@ -97,8 +20,7 @@ interface Props {
   sxIcon?: SxProps;
   sxSubheading?: SxProps;
 }
-
-const CardBase = (props: Props) => {
+export const CardBase = (props: CardBaseProps) => {
   const {
     checked,
     heading,
@@ -138,5 +60,3 @@ const CardBase = (props: Props) => {
     </CardBaseGrid>
   );
 };
-
-export default CardBase;
