+++ packages/manager/src/components/SelectionCard/SelectionCard.stories.mdx
@@ -4,7 +4,7 @@ import InsertPhoto from '@mui/icons-material/InsertPhoto';
 import DE from 'flag-icons/flags/4x3/de.svg';
 import US from 'flag-icons/flags/4x3/us.svg';
 import { Chip } from '/src/components/core/Chip';
-import SelectionCard from '/src/components/SelectionCard/SelectionCard';
+import { SelectionCard } from '/src/components/SelectionCard/SelectionCard';
 
 <Meta title="Components/Selection Card" component={SelectionCard} />
