--- packages/manager/src/components/SelectionCard/index.tsx
@@ -1,2 +1,0 @@
-import SelectionCard from './SelectionCard';
-export default SelectionCard;
