+++ packages/manager/src/features/Linodes/LinodesCreate/SelectBackupPanel.tsx
@@ -13,7 +13,7 @@ import Typography from 'src/components/core/Typography';
 import Grid from '@mui/material/Unstable_Grid2';
 import { Notice } from 'src/components/Notice/Notice';
 import RenderGuard, { RenderGuardProps } from 'src/components/RenderGuard';
-import SelectionCard from 'src/components/SelectionCard';
+import { SelectionCard } from 'src/components/SelectionCard/SelectionCard';
 import { formatDate } from 'src/utilities/formatDate';
 import {
   withProfile,
