+++ packages/manager/src/features/Linodes/LinodesCreate/SelectionCardWrapper.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import Info from 'src/assets/icons/info.svg';
 import { styled, useTheme } from '@mui/material/styles';
 import Grid from '@mui/material/Unstable_Grid2';
-import SelectionCard from 'src/components/SelectionCard';
+import { SelectionCard } from 'src/components/SelectionCard/SelectionCard';
 import { APP_ROOT } from 'src/constants';
 
 interface Props {
